==> FinalWork/1_DataPreparation_DatasetsMerge.ipynb <==
import pandas as pd

df = pd.read_csv('data/ga_hits.csv')
df.head()


print(df.info())
print(df.shape)

Целевое действие — события типа «Оставить заявку» и «Заказать звонок»
(ga_hits.event_action in ['sub_car_claim_click', 'sub_car_claim_submit_click',
'sub_open_dialog_click', 'sub_custom_question_submit_click',
'sub_call_number_click', 'sub_callback_submit_click', 'sub_submit_success',
'sub_car_request_submit_click']).

df.event_action.value_counts()

print('Количество незаполненных session_id: ',df.session_id.isna().sum())
print('Количество незаполненных event_action: ',df.event_action.isna().sum())

#Оставим только нужные столбцы
hits= df.copy()
hits = hits[['session_id', 'event_action']]
hits.head(5)

# Трансформируем столбец в удобный нам формат: 0 для нецелевых действий и 1 для целевых
def event_action_transform(event_action):
    target_list= ['sub_car_claim_click', 'sub_car_claim_submit_click',
    'sub_open_dialog_click', 'sub_custom_question_submit_click',
    'sub_call_number_click', 'sub_callback_submit_click', 'sub_submit_success',
    'sub_car_request_submit_click']
    if event_action in target_list:
        event_action = 1
    else:
        event_action = 0
    return event_action

hits['target'] = hits['event_action'].apply(event_action_transform)
hits.head()

hits_clean = hits.drop(['event_action'], axis =1)
print('Количество значений: ', hits_clean.target.value_counts())

#Сгруппируем датасет по session_id: если хоть в каком-то из событий сессии целевое действие было =1, берем 1. Иначе 0

grouped_hits = hits_clean.groupby('session_id').max()

print('Размер полученного датасета:', grouped_hits.shape)
grouped_hits.head()

grouped_hits.target.value_counts()

#Сохраним датасет
grouped_hits.to_csv('data/grouped_hits.csv')

## Датасет sessions

#Второй датасет
df2 = pd.read_csv('data/ga_sessions.csv', sep = ',', dtype = {'client_id':'string'})
df2.head()

df_all = df2.merge(grouped_hits, on='session_id', how='inner')
df_all.to_csv('data/df_all.csv')


#Для пайплайна сделаем балансировку датасета (в блокноте с анализом будем рассматривать весь датасет целиком):
df_1=df_all[df_all['target'] == 1]
df_0=df_all[df_all['target'] == 0]
df_0=df_0.iloc[:50314] 
df_balanced = pd.concat([df_1, df_0], axis=0)
df_balanced = df_balanced.sample(frac=1).reset_index(drop=True)
df_balanced.to_csv('data/df_balanced.csv')
df=pd.read_csv('data/df_balanced.csv', index_col=0)
df.head()


df.shape
